Add additional database indicies
This is a bit speculative as I have no way to test this locally (not
enough seed data), but it seems reasonable that anything that is
referenced from one table to another would benefit from an index - these
all look to meet that criteria, so I've added one.

For the record:

Testing in production with the 'Conversion funding agreement letters
export' from April 2024 which contains 116 projects.

```
Benchmark.measure do
  projects = ProjectsForExportService.new.funding_agreement_letters_projects(month: 4, year: 2024)
end
```

~2 seconds to fetch the projects and get their details from the
Academies API

```
Benchmark.measure do
  Export::Conversions::FundingAgreementLettersCsvExporterService.new(projects).call
end
```

~38 seconds to generate the csv (Rails caches after this and we drop to
~6 seconds)