Update transfer by month export view
This view is no longer consistent with the other three export index
views, we also found a bug of sorts, where the linked views in the old
layout linked to views that show conversions and not transfers.

This works brings the view inline with the others, showing a simple
count of transfers in the month and a link to the export. This solves
both issues.