Update significant_date_in_range scope
The 'by month' views show only confirmed projects i.e. those with a
significant date that is NOT provisional.

There is no reasons to deal with the SignificantDateHistory here as we
have already filtered out any project that does not have a confirmed
significant date and we only want confirmed projects anyway.

The `significant_date` field holds this date and we can query it.

We only need to deal with SignificantDateHistory when we want projects
that were due to convert or transfer but have 'slipped' or 'slid' to
another month.

As we are looking for performance improvements, I will record the
current approx time for a request/response in production now, so that we
can compare the effect:

Date Range: Jan 2024 to Dec 2024
Date run: 25 Mar 2024
Time: ~13163ms