Update FreeTDS
We rely on the Rails `activerecord-sqlserver-adapter` to talk to MS SQL
Servers, this relies on the `tiny_tds` gem, which relies on `FreeTDS`.

This work updates the currently used version of FreeTDS and makes the
build arg more conventional by switching to all capitals.

We are looking for areas where we can gain a bit of performance and
whilst no specific mention of this was found, it seems reasonable to
stay current for this vital dependency.