(Fix) Clear the value of sponsored_support_grant_type if "Not applicable" ticked
We had a bug (more a UI anomaly) on the Sponsored Support Grant task: a user
had previously selected a support grant type but now wanted the task to be
"Not applicable". However because the support grant types are in radio buttons,
they were unable to clear the value.

This confused the user, as although the task was now correctly logged as "not
applicable" the UI showed the type still selected.

Override the `save` method on these two tasks, so that if the Not applicable
box is checked, any value in sponsored_support_grant_type is removed.