Update member_for_constituency
This method returns two different objects, unlike all the other methods
for this client.

- When successful it returns the `MemberDetails` object
- When unsuccessful it returns the`Result` object (like all other
  methods in the client)

But this method is called from Project to return an MP and then we
generally guard against a nil object. In the case where this fails the
`project.member_for_constituency` is a `Result` object and not `nil` so
things go wrong further along.

The solution is to always return the `Result` object and then handle
that in the calling code.

Our first step is to return the Result with the `MemberDetails` and nil
error.

With that done, we handle the `Result` object in the calling code,
allowing a `Project` to return `nil` for `member_of_parliament` when we
cannot fetch on - which is what we want.

We tweak a couple of specs to handle this change and move on.