Move track event to Project
The members API client should really return a result with either an
error or the object that was fetched.

Here we take the responsibility for tracking any errors in the API call
out of the client and into the calling code, here in Project.

This means will get insight on all the errors coming from the API calls
and we only have one place to manage that.

We've made the actual error message easier to identify in
Application Insights by adding a prefix.

We've updated some of the stubbed errors to return objects instead of
their class, which is now required.