Add an Information banner
We've used this banner before for scheduled maintenance so it makes
sense to make it even more generic and quicker to use.

Here we add it to every page when the appropriate environment variable
is set to true and add the content to the locales to make identifying
and update quick and painless.