Add the chair of governors contact task
This tasks forces users to collect the contact details for the chair of
governors of the school converting into an academy so that a copy of the
funding agreement letters can be sent to them.

This is our second attempt at a task that collects a contacts (we
already have 'main contact') and we wanted to try something different.

Here we collect the name, email and optionally phone number and the task
is only completed one we have these. This creates a contact and
populates it with the further information that we can derive from the
project:

- organisation = school name
- role (title) = Chair of governors
- establishment urn = Project.urn

Because of the way this works, 'deleting' the contact details in the
task and saving it WILL NOT delete the contact, this must be done via
External contacts > Change > Delete contact. Updating the details of the
contact will work, however.