Extract trust exists validator
We are adding an edit form to project to allow users to alter some
project attributes. One of which is the Incoming trust UKPRN, we want to
carry out the same validation as when we are creating a new project so
we extract the custom validation out so it can be reused.

We cast the value to an integer to keep it consistent, a failure to cast
would result in the trust not existing anyway, which it could be argued
is the correct error state.

Co-authored-by: Lorna <lorna@dxw.com>