Add scoping for deleted and non-deleted projects.
Here, we add a `deleted` scope, to return projects which have
`deleted` set to true.

We considered adding the opposite scope, `active` for non-deleted projects.

However, we don't want to ever show deleted projects in the UI, in exports or
in statistics. Adding an `active` scope (for non-deleted projects) to all the
places where we gather projects will be a lot of work. Instead, use
`default_scope` to always return non-deleted projects (active projects) in the
application.

We are using `where.not(deleted: true)` in the default scope to catch any
instances where `deleted` may be false OR nil.

In the rare instance we want to show deleted projects, we can explicitly call
the `deleted` scope.