(Fix) Fix a subtle bug with assigning the Chair of Governors contact
Adding the `default_scope` to Projects uncovered a very odd anomaly with this
test for updating an existing Chair of Governors contact.

Line 362 here was creating a contact to act as chair of governors, but because
the contact was created but NOT assigned to the project, when we tried to update
the Chair of Governors the test didn't work. I am not sure why/how adding the
`default_scope` to Projects has caused this to happen.

When the `chair_of_governors` object is created, it does not have the
`project_id` of the project we are assigning it to as the chair of governors on
line 363. But when we try to update it later in the test, we try to change the
`project_id` to the ID of `project` (i.e. changing its assigned project) and
the test fails.

Explicitly assigning the `chair_of_governors` object to the `project` on
creation makes the test pass.

We have also added an extra expectation to ensure the chair of governors contact
is updated.