Service support users can "soft delete" a project
Now that projects can have a "deleted" state (where they do not appear on the
UI but are still in the database) we want to allow service support users to
"soft delete" a project.

To do this, we have used a "delete" route as opposed to the usual "destroy". As
a result, the delete routes are not the same as the destroy routes for notes
and contacts. "Destroy" implies the record is removed from the database, while
the projects are only hidden, not removed.

We use a "confirm delete" pattern to be absolutely sure the user wants to
delete the project, as there is currently no way for a user to restore a
"deleted" project.

Once a project is "soft deleted", the user is returned to the All Projects
list page, with a confirmation the project was deleted.