Introduce a concept of "state" to projects
We want to be able to mark projects as "deleted", but not actually delete
the record from the database. We intend to add a deleted state to projects.

Ahead of this, we have decided to add a `state` to projects, and enumerate it
for the various different states a project can be in. For now, this is "active"
and "completed"

Add an enumerator to projects - 0 for active (default) and 1 for completed.

As the state is an enumerator, we get the scope Project.completed for free now.
Due to Rails magic, we also get Project.not_completed - anything not in the
completed state. For now, this is "active", but in future if we add more states
we will need to make "not complete" more specific (ie if we add a deleted
state, then deleted will also fall under "not completed")

As the original Project.completed scope had an order (ordering by completed_at
date) we have had to split this ordering out and add it as a new scope.

We have added the new state to the project factories, and updated the relevant
tests.

Unfortunately this was the smallest single commit we could make for this
change.