Add a GOVUK Date field parameter validation class
If you didn't know, the GOVUK Formbuilder date handling is a mystery (to
me at least!). We have at least three existing ways of giving users
meaningful feedback on the values they put into the three form inputs.

I have no idea why other projects do not suffer from this, I can only
assume they do not worry about giving users feedback and simply let the
`Date` be set as nil and tell the user the date is missing. That's not
good enough for us, so here we are, working around!

I suspect another approach is to not use the Formbuilder date field and
treat each value separately, validating them accordingly and mashing
them back into a Date object - but for now here is another approach!

Let's describe the process as it would work for context:

In more usual Rails MVC for an edit we would load the model that is
being edited and display a form with the current values.

Users would update the values in that form and POST to the update
action of the controller, generally we would then mass assign the posted
parameter - passing them through the strong parameter process first and
then attempt to save the model with the new values.

If the model was invalid for any reason we would render the edit form
again and show the users any validation errors on the model along with
the posted values.

Now the three field values of a date object get sent as super special
'multiparameter' values (these are the `(3i)` bits you will see) and in
our normal context get cast to a date during the mass assignment
(`asssign_attributes`). We would usually see this taking place in the
controller and indeed if you force bad date params in you will raise a
`ActiveRecord::MultiparameterAssignmentErrors` in the controller.

Now the GovukDateFieldParameters job is to check over the parameters
BEFORE they get mass assigned and so prevent the any errors from being
raised. You can check the specs and comments for what we look out for,
unfortunately Ruby Dates can be created via all number of input that
users of our application would not recognise as making a more
conventional date and could cause incorrect dates to be successfully
saved.

The class cannot actual prevent the assignment, that is up to the
developer, it simply returns true of false to valid? or invalid?

Once a developer knows if the parameters are valid or not they can take
the appropriate action, usually this would be to save a model or add
validation errors to the model for display, it might look something like:

```
if GovukDateFieldParameters.new(:date_attribute, params).invalid?
  errors.add(:date_attribute, :invalid)
  return false
end

assign_attributes(params)
```

This probably feels a lot like a traditional Rails validator and you
might be wondering why we don't write one. The reason is that the three
parameters are used BEFORE any validation can take place, in the mass
assignment, so a validator does us no good.

This also means we have to move the assignment down into the model so we
can act on bad parameters and inform the user. For us, our models are
form objects so this works out well, with the form object updating any
actual models internally.

I think there is enough flexibility here that you can use this wherever
you want to, but for us, it will be in form objects that are editing a
Date attribute.

We'll use this class on an actual form in the next commit and possibly
refactor existing cases of Date attributes to use this as well.

Sorry for the giant commit message and please, if you know better, let
me know! I am so sick of having to come up with workarounds for this!