Users can change the advisory board details
As the advisory board includes a date we leverage the new
GovukDateFieldParameters class which means altering how we handle this
edit form object.

We first initalize a new form object from a project, this is then used
to render the edit form.

A user modifies the values and posts those to the `update` action which
again loads a new form object from the project, but then we call the
`update` method passing along the strong parameters, this let's us
validate that the parameters for the date are 'valid' and take any
actions if they are not, this usually means showing a form error.

I am hoping this feels like the way a regular model would get updated,
the main difference being we have to do the mass assignment in the form
object rather than the controller so the errors can be handled.

When the date parameters are valid we go ahead and let the mass
assignment take place, this is when Rails turns the three separate
values into a single Date object.