(Fix) A contact must be assigned to a project before it is added as chair of governors
While writing code elsewhere we uncovered a small anomaly in this test - the
contact which is created to be the Chair of Governors contact for a project
must actually belong to that project before it is assigned. In not linking the
contact to the project, the test is actually reassigning the contact from one
project to another, as well as updating its attributes.

This anomaly was showing up as a bug in other branches (specifically when
working on soft deleting projects).