Do not return deleted projects by default (`default_scope`)
Adding default_scope to models can be problematic as default scopes can conceal
issues. However, now that we have a concept of "soft deleted" projects, we do
not want to return these projects ANYWHERE in the application - not in the UI,
exports or statistics. They should not be seen to exist.

To avoid explicitly asking for `active` and/or `completed` projects in all
our pages, use `default_scope` to exclude deleted projects.

If we want to view deleted projects in future, we will have to explicitly look
for them (e.g. using `Project.unscoped.where(state: :deleted)` as opposed to the
Project.deleted scope).