Use ProjectPolicy to stop individual deleted projects being viewed & edited
We don't want users viewing deleted projects by going to their `show` URL - so
ensure this doesn't happen using Policies.

This will return a 403 not a 404, which semantically isn't quite correct - the
project isn't deleted, so it's "found" but it's not viewable. But the user
shouldn't know it's not "deleted". However we think it will be rare for a user
to have a project URL and go to it directly.

The editing restricton is more of a theoretical need than a practical one, but
mark "soft deleted" projects as un-editable, just in case anyone manages to
access one and tries to edit it.