Expose a based project date history
We store and some users want to see the history of the changes to the
significant date over time.

This commit adds a `/dates` path to projects and displays a table of the
date history.

We have intentionally not added a link to this view as it has not been
designed, it is merely a technical response to the fact we model this
behaviour.