Add a New trust name validator
We want users to use the same New trust name on all projects in a 'form
a multi academy trust' collection.

To aid this we want to validate that:

- if a project for the same new trust already exists, the provided trust
  names must match

The validator relies on the first returned project i.e. the first
project in the application locks the trust name, subsequent projects
will have to match or the first be corrected.

The trust names are not case sensitive, we convert them to UPPERCASE for
comparison.