Add reason to RPA task
When an academy decides to use commercial risk protection arrangement we
want to know why. This work adds a attribute and text field to collect
the reason IF the value is 'commercial'.

We validate the presence of the reason when the `option` is 'commercial'
and we only consider the task complete with `reason` present when `option`
is 'commercial'.

We decide to store `reason` rather than add a note as we felt the
context was different with the reason only supplied under specific
conditions.

Requiring the reason sets this task apart as you can become 'stuck' on
this task but accidentally choosing a value when you didn't mean to, but
given that we want to enroll academies in the 'standard' option when
this task is incomplete, this is probably fine.