Fix the RPA reason export value
In the export presenter the `option` was assigned the reason value when
it should be the option value.

This issue presented itself when we viewed existing projects in
development that were already set to 'commercial' but had `nil` reason -
this case is an exception we cannot do anything about as we don't know
the reason, the only option is for users to backfill the values.

Any project that has RPA of 'commercial' will now be in progress until a
reason is given.