All projects in progress becomes tabbed
We want to show different scopes of in progress projects with tabs.

This is the setup for that by changing the existing 'all projects in
progress' to be inside a tab.

I don't like the title being outside of the tab scope in the view, but
this is what the design calls for and we have other instances of this so
I am going with it.

To make all the things work we need to add a component to the existing
path, which now means something like:

> projects from all teams, that are in progress, that are conversions
and transfers (all types)

We could have added this to the existing routes block, but the id
constraints are not required for these index only routes and make
debugging routes harder, so we settled on declaring this new route in a
new block.

All this leads us to the unfortunate path helper:

`all_all_in_progress_projects_path`

which means:

> all project types (conversions and transfers) for all teams that are
in progress

which is accurate, but looks odd. We did try to make this nicer, but
could not find a decent way. We hope that once we have others, like:

`conversions_all_in_progress_projects_path`

this will become more approachable.