Add a Form a multi academy trust project group
This object provides a simple interface to a collection of projects with
the same `new_trust_reference_number`, known as a 'Form a MAT project'.

Whilst it is not a project in itself, users think of it along the same
lines and so we authorise it as a project, doing so means it must
implement `deleted?` for the policy to work correctly.

When no projects can be found we raise an error that can then be handle
appropriately.