Add a service to fetch all form a MAT project groups
We already have a model for a form a multi academy trust project group
and now we want users to be able to see a list of all of them.

We first create a service object to handle this. It fetches all active
projects and collates the `new_trust_reference_numbers`.

It then removes any duplicates and nil values leaving a distinct list of
trust reference numbers.

This list is mapped over and new FormAMultiAcademyTrust::ProjectGroup
created for each.

When no matching projects can be found, we return an empty array so that
the calling code can take the appropriate action.