Update Form a MAT Project Group to include region
The UI designs call for this although it is even more brittle than
having to fetch the name from the first project, we have no way to store
the region against the 'form a multi academy trust' so this is all we
can do.

If a collection of projects does happen to end up with different regions
this could cause confusion.