Add assigned to to ProjectGroup
The design for the form a multi academy trust index view calls for the
assigned to user, but we have no concept of that, the group is a
collection of projects, each of which is assigned to a user.

As we already select the region of the first project, we've done the
same here, returning the user assigned to the first project in the
collection which is also the first project added and so also the one
that sets the trust name.