Add a from a MAT index view
This view allows users to view all form a multi academy trust projects
grouped together.

The ProjectGroupFetcherService does the work to collate the projects
together and we display them in a table.

One issue we encountered was a Bullet false positive, we've ended up
disabling Bullet on the spec and left a comment, but to summarise:

The object fetches a group of projects with the same TRN and is designed
to represent a SINGLE collection, generally when we view a single
collection we show each project and which user it is assigned to so the
service eager loads the assigned_to Users. i.e the internal set of
projects has its associated assigned_to users.

When the service creates multiple objects to show all the collections,
we only used the first assigned_to User so it still feels worth eager
loading. However we pass each collection of projects off to the
projects_establishment_name_list helper which does NOT use the eager
loaded assigned_to and so bullet raises the AVOID eager loading warning,
but the code outside the helper will use one of the assigned_to users.

The alternative is to add complexity and rely on the developer to
know when to eager load in the collection objects - which felt worse
than disabling Bullet for one spec.