Make the "Date academy opened?" task mandatory to complete a project
A project should only be completed if the date the academy opened is present.

Alter the `completable?` method to add this as another mandatory task.

Also add a note to the error message shown when a project is not "completable",
to let the user know they must complete all mandatory tasks.