Remove region and assigned
Because form a multi academy trust project groups do not exists, we just
load a collection based on the 'trust reference number', we cannot
really show the region and assigned to for the collection, they could be
different.

We did show them by using the values from the first project in the
collection, but this could cause confusion so we decided to take the
columns out.