Display school type instead of phase if school phase is not applicable
If a school is a PRU/special school, the school phase will be
"Not applicable" within the Academies API.
ESFA do not want to see "Not applicable" in the Phase column within the
exports. So instead we now show the school type in the phase column if the phase is
missing.