Refactor out the previous `Project.completed?` method
We previously determined if a project was completed by seeing if the
`completed_at` date was present. Since we added an enum value for Project.state,
which includes `completed`, we get `Project.completed?` for free. The enum
method looks to see if a project has a state of `completed`.

We forgot to remove the old `Project.completed?` method which looks at the
value of `completed_at`. This was an oversight.

Remove this old method and let the enum value of `state` take over. When we
introduced the enum, we did a data migration to update all projects with a
`completed_at` date to have the `completed` state, so this refactor should not
affect the live data.