Rename the Grant management routes/controllers to "By advisory board date"
The exports part of the app has become quite confusingly named. We originally
named the exports after who would be consuming them (the Grant management team,
ESFA etc) but after some thought we've decided that it makes more sense to name
them after the way the projects are grouped. The Grant management and finance
exports are grouped by the project's advisory board date, so rename the routes
and controllers in this way.