Refactor the remaining "By month" exports into BySignificantDate namespace
We had "By month" exports in two different places, doing two different things.

In a previous commit we refactored the "By month" exports which were used in the
"By month" part of the UI, into a new BySignificantDate namespace.

We now refactor the "By month" exports which were in the Exports section, into
this new namespace as well. These are the "all in one" exports for ESFA,SUG and
FA letters, and are named "Academies due to transfer" and "Schools due to
convert"