Change project user assignment
We currently progressively enhance the user assignment form from a
select to a autocomplete. Whilst this is great for small numbers of
items, we have close to 500 users in production now and so we should stop
potentially showing a non JavaScript user a select with 500 entries.

To use a full autocomplete, we first have to ensure that non JavaScript
users can still complete the task. Here the task is to assign a user to
a project, so we switch out the select for a text field that accepts the
users email address.

We validate the email address is `@education.gov.uk` and is a user that
can be assigned to a project via the `assignable` scope.

We take the opportunity to update the paths to be more conventional and
move the sending of the email into the form object, thinning out the
controller.

We also fix the dynamic redirect by setting a path in the session and
returning the user to that path upon submission of the form or
cancelling (this previously only worked on form submission).

At this point there is no autocomplete, so the user experience has been
degraded, but we will add that in later work.