Remove the "By month" view from the Team section
Under the "Team" section we had a By month view, which was divided into Revised
and Confirmed project (by date). This section hadn't actually worked for a while
- it was broken when we implemented the new By month section under All projects.
We did not have adqeuate test coverage to show it was broken.

Because no users had complained about it being broken, we assume the section
is not being used by users, and have decided to remove it.