Add a user search endpoint
This endpoint will allow the searching of users and will be used by the
new assignment autocomplete that we are working on.

The endpoint is designed to work with the autocomplete which formats a
user like this:

`firstname lastname (firstname.lastname@education.gov.uk)`

We've added a `type` that can be used later to restrict the types of
users that the auto complete will use.

As this is for autocompletion, we always return a result, even if it is
empty.