Replace ESLint with Standard and update linting
We use standardrb so why not use standardjs as well - it's simpler to
maintain and make lots of decisions for us.

We have also switched to an explicit approach, only linting the
things  we write, rather than linting the entire project and having
to exclude lots of paths. This has the added benefit of being much
faster and the cost of missing something later, which is pretty low as
we are only linting.

We have also moved the dfe frontend to the vendor directory, it should
really be a package, but the package is broken, so this is the best
place for it.