Switch to JS Bundling
We will need some modern JavaScript so we need to install and setup JS
Bundling[1]. We've chosen ESBuild as it is quick and simple to setup,
allowing us to target our bundles at browser version.

This change does mean developers will need to run the app with:

`bin/dev`

from here to ensure the JS is bundled, the developer scripts have been
update so the impact should be minimal.

Since we moved to JSBundling we have to ensure the JS bundles are in-place
before any tests can take place.

For all but our accessibility tests, this is an simple as switching to
the Rails tasks, as these run the JS builds.

Our accessibility tests are more complicated, because Rspec excludes
them by their tags by default. We need to run Rspec directly so we can
passs `--tag` which means we can't use the Rails tasks, which means we
have to ensure the JS is bundled ourselves.

We can do:

`bin/rails javascript:build && bin/rspec --tag accessibility`

But we have to be in Bash, so we change the Docker `run` command to suit.

We went ahead and changed all the CI to use this approach for
consistency.

[1] https://github.com/rails/jsbundling-rails