Add autcomplete to edit assigned user form
With this form updated we can progressively enhance it with the
accessible autocomplete.

We kept seeing flaky specs that seem to relate to the session, we would
not be able to locate the User with the id from the session. We first
added a sign out after each spec but still had the issue, adding the
sleep at the end of the sign out is the only thing that has worked.