Update edit added by user
Like the assigned user we want to drop the select box that backs the
autocomplete on this edit form, to do so we switch it for a text field
that takes the email address of the user, validating the choice.

We drop a spec here as we have coverage for this behaviour in the two
new specs for the edit added by and assigned to.