Conversion stakeholder kick off task date validation
We want all out handling of date field parameters to be as similar as is
possible.

This works update the conversion project stakeholder kick off task,
which collects the confirmed conversion date.

Our approach is to use the `GovukDateFieldParameters` class to check the
parameters before they are assigned and cast to a Date object.

Tasks are a little bit different context as they assign the attributes
in the controller, so we override the `assign_attributes` method.