Update GovukDateFieldParameters
This class is tasked with taking submitted date parameters and checking
they are in acceptable ranges.

We accept dates without a day, which actually gets submitted as "1", we
need the class to handle that.

We have no way to know if the submitted dates are from a date field
without a day, so we have to make it an option that the developer can
set.

With this in place we can treat parameters without a day as an empty
submission, which the class is designed to consider valid.