Update change significant date validation
We are consolidating our approach to date validation to use the
`GovukDateFieldParameters` class before assignment.

This work updates the conversion and transfer change significant date
form to use it. Because the validation sets any invalid dates to `nil`
and for this form, nil is not allowed, we can do nothing other than show
the error for blank state.