Service object to create Conversion::Projects with the bare minimum attributes
We are going to create Conversion projects via the API, from data stored in the
Prepare application.

Because Prepare does not store all the data points we need, we are going to
create projects using the most "bare bones" attributes we can. In order to do
this we must save the project with validation turned off.

We also have to check to see if the user associated with this project in
Prepare exists on Complete. If they do not, try to create a new user (put them
in the RCS team for now) and associate them with the project.