Install and configure Swagger UI
Now that we have Grape generating documentation, we can expose it with
Swagger UI [1].

We have chosen to use the 'dist' method [2] to run the UI as we don't want
to load up our js and css bundles with code only used by the API
docs.

We load the dependency via Yarn and add it to the asset pipeline. We
create a new layout, controller and view to render the UI at
`/api/docs`, the layout allows us to load all the assets from the dist
package.

We have to mess with the content security policy and unfortunately, we
have to rely on `unsafe-inline` to support the Swagger UI - this seems
to be an issue with the UI not passing a nounce down to the scripts and
styles it loads, we could not get this to work any other way!

We only allow the `unsafe-inline` for `style-src-elem` [3] and only on the
API docs themselves to minimise any risk.

We also have to disable StandardJS from liniting the
`swagger-intializer` as it is using classes that are only available at
runtime - we couldn't find a way around this without significant work.

The route we use is publicly available as we think the users of the API
are not the same as those that use the application.

[1] https://swagger.io/tools/swagger-ui/A
[2] https://www.npmjs.com/package/swagger-ui-dist
[3] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/style-src-elem