Use validation on URN and UKPRN in the API project creation service
We use these validators to validate URN and UKPRN in the Create project form and
Project model. Use them in the service object that creates projects via the
API as well.

We also had to correct the URN in one of the other tests, as the original URN
in that test (12345) was invalid. That test was not demonstrating param
validation failure, so correct it to get the right test case!