DRY up params for both Conversion project endpoints
The 'regular' Conversion project and Form a MAT Conversion project endpoints
both have very similar required parameters. We can DRY up the parameters by
sharing the parameter sets between both methods (excluding the params which
are unique to either endpoint).

This means we've lost the `conversion_project` param wrapper but on reflection
it was not needed as we know what "type of thing" we are creating by the endpoint.