Change the "All conditions met" task for transfers
The "All conditions met" task should be called "Authority to proceed" for
Transfers.

We have decided to keep the name of the attribute the same (`all_conditions_met`)
for simplicity, but change the accompanying text. We have added an alias
`authority_to_proceed` on the attribute.

We have also added two more checkboxes to this task. In a future commit we will
do a data update to fill these checkboxes for completed transfer tasks.